requirements.txt
Flask==1.1.2
pycryptodome
scrypt
paho-mqtt
cryptography
